til 인증 validation 추가

+++ src/main/java/com/debrains/debrainsApi/controller/TilCrtController.java
@@ -4,15 +4,20 @@ import com.debrains.debrainsApi.dto.TilCrtDTO;
 import com.debrains.debrainsApi.entity.TilCrt;
 import com.debrains.debrainsApi.exception.ApiException;
 import com.debrains.debrainsApi.exception.ErrorCode;
+import com.debrains.debrainsApi.hateoas.TilCrtConverter;
 import com.debrains.debrainsApi.repository.TilCrtRepository;
 import com.debrains.debrainsApi.security.CurrentUser;
 import com.debrains.debrainsApi.security.CustomUserDetails;
 import com.debrains.debrainsApi.service.TilCrtService;
 import com.debrains.debrainsApi.util.PagedModelUtil;
+import com.debrains.debrainsApi.validator.TilCrtValidator;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
+import org.modelmapper.ModelMapper;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.data.web.PageableDefault;
 import org.springframework.data.web.PagedResourcesAssembler;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.hateoas.Link;
@@ -26,6 +31,9 @@ import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 import java.net.URI;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.LocalTime;
 
 import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
 
@@ -39,13 +47,26 @@ public class TilCrtController {
 
     private final TilCrtRepository tilCrtRepository;
 
+    private final TilCrtValidator tilCrtValidator;
+
     /**
      * til 인증 생성
      */
     @PostMapping
-    public ResponseEntity createTilCrts(@RequestPart(value = "files", required = false) MultipartFile[] files,
+    public ResponseEntity createTilCrts(@CurrentUser CustomUserDetails currentUser,
+                                        @RequestPart(value = "files", required = false) MultipartFile[] files,
                                         @RequestPart(value = "tilCrtDTO") @Validated TilCrtDTO tilCrtDTO)
             throws IOException {
+        tilCrtValidator.validateDate(tilCrtDTO);
+        tilCrtValidator.validateCrt(tilCrtDTO, files);
+
+        LocalDateTime start = LocalDateTime.of(LocalDate.now(), LocalTime.MIN);
+        LocalDateTime end = LocalDateTime.of(LocalDate.now(), LocalTime.MAX);
+
+        Long tilCrtCount = tilCrtRepository.tilCrtCount(currentUser.getId(), start, end);
+        if (tilCrtCount > 0) {
+            throw new ApiException(ErrorCode.TILCRT_TODAY);
+        }
 
         TilCrt newTilCrt = tilCrtService.createTilCrts(files, tilCrtDTO);
         var selfLinkBuilder = linkTo(TilCrtController.class).slash(newTilCrt.getId());
@@ -53,8 +74,8 @@ public class TilCrtController {
         EntityModel<TilCrt> resource = EntityModel.of(newTilCrt);
         URI createdUri = selfLinkBuilder.toUri();
         resource.add(linkTo(TilCrtController.class).slash(newTilCrt.getId()).withSelfRel());
-        resource.add(linkTo(TilCrtController.class).withRel("query-til-crts"));
-        resource.add(selfLinkBuilder.withRel("update-til-crt"));
+        resource.add(linkTo(TilCrtController.class).withRel("til-crts"));
+        resource.add(selfLinkBuilder.withRel("update"));
         resource.add(Link.of("/docs/index.html#resources-til-crts-create").withRel("profile"));
 
         return ResponseEntity.created(createdUri).body(resource);
@@ -64,14 +85,29 @@ public class TilCrtController {
      * til 인증 리스트 조회
      */
     @GetMapping
-    public ResponseEntity<PagedModel<EntityModel<TilCrt>>> queryTilCrts(Pageable pageable,
-                                                                        PagedResourcesAssembler<TilCrt> assembler) {
+    /*public ResponseEntity queryTilCrts() {
+
+        List<TilCrtDTO> page = tilCrtRepository.findAll()
+                .stream().map(entity -> modelMapper.map(entity, TilCrtDTO.class))
+                .collect(Collectors.toList());
+
+        List<EntityModel<TilCrtDTO>> tilCrtList = page.stream()
+                .map(tilCrtDTO -> tilCrtConverter.toModel(tilCrtDTO)).collect(Collectors.toList());
+
+        CollectionModel<EntityModel<TilCrtDTO>> resource = CollectionModel.of(tilCrtList,
+                linkTo(methodOn(this.getClass()).queryTilCrts()).withSelfRel());
+
+        return ResponseEntity.ok(resource);
+    }*/
+    public ResponseEntity<PagedModel<EntityModel<TilCrt>>> queryTilCrts(
+            @PageableDefault(sort = "id", direction = Sort.Direction.DESC) Pageable pageable,
+            PagedResourcesAssembler<TilCrt> assembler) {
 
         Page<TilCrt> page = tilCrtRepository.findAll(pageable);
 
         PagedModel<EntityModel<TilCrt>> resource = PagedModelUtil.getEntityModels(assembler, page,
                 linkTo(TilCrtController.class), TilCrt::getId);
-        resource.add(Link.of("/docs/index.html#resources-tils-list").withRel("profile"));
+        resource.add(Link.of("/docs/index.html#resources-til-crts-list").withRel("profile"));
 
         return ResponseEntity.ok(resource);
     }
@@ -80,14 +116,17 @@ public class TilCrtController {
      * til 인증 상세 조회
      */
     @GetMapping("/{id}")
-    public ResponseEntity getTilCrt(@PathVariable Long id) {
+    public ResponseEntity getTilCrt(@CurrentUser CustomUserDetails currentUser, @PathVariable Long id) {
 
         TilCrt tilCrt = tilCrtRepository.findById(id)
                 .orElseThrow(() -> new ApiException(ErrorCode.TILCRT_NOT_FOUND));
 
         EntityModel<TilCrt> resource = EntityModel.of(tilCrt);
-        resource.add(linkTo(TilCrtController.class).withSelfRel());
+        resource.add(linkTo(TilCrtController.class).slash(tilCrt.getId()).withSelfRel());
         resource.add(Link.of("/docs/index.html#resources-til-crts-get").withRel("profile"));
+        if (tilCrt.getUser().getId() == currentUser.getId()) {
+            resource.add(linkTo(TilCrtController.class).slash(tilCrt.getId()).withRel("update"));
+        }
         return ResponseEntity.ok(resource);
     }
 
@@ -95,10 +134,13 @@ public class TilCrtController {
      * til 인증 수정
      */
     @PatchMapping("/{id}")
-    public ResponseEntity updateTilCrt(@PathVariable Long id,
+    public ResponseEntity updateTilCrt(@CurrentUser CustomUserDetails currentUser,
+                                       @PathVariable Long id,
                                        @RequestPart(value = "files", required = false) MultipartFile[] files,
                                        @RequestPart(value = "tilCrtDTO") @Validated TilCrtDTO tilCrtDTO) throws IOException {
-
+        if (!currentUser.getId().equals(tilCrtDTO.getUserId())) {
+            throw new ApiException(ErrorCode.NO_AUTHORIZATION);
+        }
         TilCrt savedTilCrt = tilCrtService.updateTilCrt(id, files, tilCrtDTO);
 
         EntityModel<TilCrt> resource = EntityModel.of(savedTilCrt);
@@ -112,16 +154,18 @@ public class TilCrtController {
      * til 인증 삭제
      */
     @DeleteMapping("/{id}")
-    public ResponseEntity deleteTilCrt(@PathVariable Long id) {
+    public ResponseEntity deleteTilCrt(@CurrentUser CustomUserDetails currentUser,
+                                       @PathVariable Long id) {
         TilCrt tilCrt = tilCrtRepository.findById(id)
                 .orElseThrow(() -> new ApiException(ErrorCode.TILCRT_NOT_FOUND));
 
-        tilCrtService.deleteTilCrt(tilCrt);
-        EntityModel<TilCrt> resource = EntityModel.of(tilCrt);
-        resource.add(linkTo(TilCrtController.class).withSelfRel());
-        resource.add(Link.of("/docs/index.html#resources-til-crts-delete").withRel("profile"));
+        if (!currentUser.getId().equals(tilCrt.getUser().getId())) {
+            throw new ApiException(ErrorCode.NO_AUTHORIZATION);
+        }
 
-        return ResponseEntity.ok(resource);
+        tilCrtService.deleteTilCrt(tilCrt);
+
+        return ResponseEntity.noContent().build();
     }
 
     @DeleteMapping("/{id}/file")
